Refactored cache location, added config env variables
