+++ app/graphql/users/middlewares.js
@@ -1,6 +1,5 @@
 const errors = require('../../errors');
 const { EMAIL_REGEX, PASSWORD_REGEX } = require('../../helpers/constants');
-const cacheHelpers = require('../../helpers/cache');
 
 const createUser = (resolve, root, args) => {
   const { user } = args;
@@ -26,17 +25,8 @@ const createUser = (resolve, root, args) => {
   return resolve(root, args);
 };
 
-// eslint-disable-next-line max-params
-const cache = (resolve, root, args, context, info) =>
-  cacheHelpers
-    .shearchInCacheOr(root, args, context, info)
-    .then(cached => (cached ? cached : resolve(root, args, context, info)));
-
 module.exports = {
   mutations: {
     createUser
-  },
-  queries: {
-    cache
   }
 };
