+++ app/graphql/users/queries.js
@@ -5,17 +5,12 @@ const resolvers = require('./resolvers');
 module.exports = {
   queries: {
     user: resolvers.getUser,
-    users: resolvers.getUsers,
-    cache: (root, args, context) => {
-      context.cache.set(context.cacheKey, 'world', 3);
-      return 'ama cache';
-    }
+    users: resolvers.getUsers
   },
   schema: gql`
     extend type Query {
       user(id: ID, name: String, email: String): User!
       users: [User]
-      cache(id: ID): String!
     }
   `
 };
